Move the pointInfo function to app.svg
