+++ [library]/SVGTurtle.js
@@ -3,7 +3,7 @@
 
 
 import { Turtle } from "./Turtle.js";
-import * as Maths from "./Maths.js";
+
 
 
 export class SVGTurtle {
@@ -78,72 +78,7 @@ export class SVGTurtle {
 		return result;
 	}
 
-	/* pointInfo
-	*/
-	pointInfo(x, y) {
-		const result = {
-			cartesian: '',
-			polar : '',
-		};
 
-		const coords = new this.turtle.space.CartesianCoordinates(x,y);
-
-		const point = this.turtle.space.newPoint('point marker');
-		point.cartesian = coords;
-
-		const svgX = x;
-		const svgY = -y;
-
-		const cartesianInfo = [
-			`x: ${point.x.toPrecision(this.precision.report)}`,
-			`y: ${point.y.toPrecision(this.precision.report)}`,
-		];
-
-		const polarInfo = [
-			`r: ${point.radius.toPrecision(this.precision.report)}`,
-			`a: ${point.angle.degrees.toPrecision(this.precision.report)}°`,
-			`a: ${point.angle.radians.toPrecision(this.precision.report)} rad`,
-			`a: ${point.angle.radiansPi.toPrecision(this.precision.report)} π rad`,
-			`a: ${point.angle.radiansTau.toPrecision(this.precision.report)} τ rad`,
-		];
-
-		const pointReport = [
-			`cartesian:`,
-			`	${cartesianInfo.join('\n\t')}`,
-			`polar:`,
-			`	${polarInfo.join('\n\t')}`,
-			`svg:`,
-			`	x: ${svgX.toPrecision(this.precision.report)}`,
-			`	y: ${svgY.toPrecision(this.precision.report)}`,
-		].join('\n');
-
-
-		result.cartesian = `
-			<line x1="${svgX}" y1="0" x2="${svgX}" y2="${svgY}"><title>${cartesianInfo.join('\n')}</title></line>
-			<line x1="0" y1="${svgY}" x2="${svgX}" y2="${svgY}"><title>${cartesianInfo.join('\n')}</title></line>
-			<use href="#def-point" class="use-point" x="${svgX}" y="${svgY}">
-				<title>${pointReport}</title>
-			</use>
-		`;
-
-		const angle180 = Maths.degrees180(point.angle.degrees);
-		const sweepFlag = (angle180 >= 0) ? 1 : 0;
-
-		result.polar = `
-			<line x1="0" y1="0" x2="${point.x}" y2="${-point.y}"><title>${polarInfo.join('\n')}</title></line>
-			<circle r="${point.radius}">
-				<title>r: ${point.radius.toPrecision(this.precision.report)}</title>
-			</circle>
-			<path d="M 0,${-point.radius} A ${point.radius},${point.radius} 0 0 ${sweepFlag} ${svgX},${svgY}">
-				<title>${polarInfo.join('\n')}</title>
-			</path>
-			<use href="#def-point" class="use-point" x="${svgX}" y="${svgY}">
-				<title>${pointReport}</title>
-			</use>
-		`;
-
-		return result;
-	}/* pointInfo */
 
 
 	moveToXY(x,y) {
